[Feature #116191173] Run successful tests
