+++ codango/static/js/tests/about-test.jsx
@@ -4,11 +4,12 @@ let FlipCard = require("react-flipcard");
 import { shallow , mount, render } from 'enzyme';
 // About-us-test.jsx
 import About from '../components/about.jsx';
+import Card from '../components/about-card.jsx';
 
 describe('<About/>', () => {
     it("it tests the about-us page elements", () => {
         const wrapper = shallow(<About/>);
         expect(wrapper.contains(FlipCard)).toBe(true);
-        expect(wrapper.find(FlipCard).length).toBe(3);
+        expect(wrapper.find(Card).length).toBe(3);
     });
 });
